New Device Dialog: Fixed Bug: Dialog not working after other popups.
The new Device Dialog wasn't working after viewing the device
page. Looks like the jquery selector for newDeviceDialog
wasn't finding it anymore after a close. Now we are only using
the selector once to reference the object, and we use that object
reference to complete any future operations for that object.

Signed-off-by: Isaac Assegai <Isaac.a.travers@gmail.com>